Removed ExploreContainer, added necessary style for posts item within the timeline stylesheet, made function in slideUpMenu return void as expected similar to ShareModal with the MenuToggleFunction, moved isOwner logic to parent component of post and passed down as a prop to post container. added expected isOwner type for prop
